admin-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
python-dotenv==1.0.1
pydantic[email]==2.10.3
pydantic-settings==2.7.0
reportlab==4.1.0
python-dateutil==2.9.0
